Rekam Medis udah nice nich fetch nya

+++ api/app/Http/Controllers/Api/Outpatient/OutpatientVisitController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Api\Outpatient;
 
+use App\Http\Controllers\Controller;
 use App\Http\Requests\OutpatientVisitRequest;
 use App\Models\OutpatientVisit;
 use App\Services\OutpatientVisit\Service\OutpatientVisitService;
@@ -10,7 +11,7 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Throwable;
 
-class OutpatientVisitController
+class OutpatientVisitController extends Controller
 {
     use ApiResponse;
 
@@ -23,6 +24,8 @@ class OutpatientVisitController
 
     public function index(Request $request): JsonResponse
     {
+        $this->authorize('view', OutpatientVisit::class);
+
         $data = $this->outpatientVisitService->getOutpatientVisits($request);
         return response()->json($data);
     }
@@ -33,6 +36,8 @@ class OutpatientVisitController
      */
     public function store(OutpatientVisitRequest $request)
     {
+        $this->authorize('create', OutpatientVisit::class);
+
         $data = $this->outpatientVisitService->store($request);
         return $this->successResponse($data, 'Outpatient visit created successfully.');
     }
@@ -40,6 +45,8 @@ class OutpatientVisitController
 
     public function update(OutpatientVisitRequest $request, OutpatientVisit $outpatientVisit)
     {
+        $this->authorize('update', OutpatientVisit::class);
+
         $data = $this->outpatientVisitService->update($request, $outpatientVisit);
         return $this->successResponse($data, 'Outpatient visit updated successfully.');
     }
@@ -47,6 +54,9 @@ class OutpatientVisitController
 
     public function show(OutpatientVisit $outpatientVisit)
     {
+
+        $this->authorize('show', OutpatientVisit::class);
+
         $outpatientVisit->load('patient');
         return response()->json($outpatientVisit);
     }
@@ -54,6 +64,9 @@ class OutpatientVisitController
 
     public function destroy(OutpatientVisit $outpatientVisit)
     {
+
+        $this->authorize('delete', OutpatientVisit::class);
+
         $data = $this->outpatientVisitService->destroy($outpatientVisit);
         return $this->successResponse($data, 'Outpatient visit deleted successfully.');
     }
